+++ zodiac_rendering_glium/src/primitives.rs
@@ -8,7 +8,7 @@ pub struct RenderPrimitive {
     pub outer_colour: [f32; 4],
     pub identification: [u16; 2],
     pub extra_data_1: [f32; 4],
-    pub extra_data_2: [f32; 4]
+    pub extra_data_2: [u16; 4]
 }
 
 implement_vertex!(
@@ -30,7 +30,7 @@ impl RenderPrimitive {
         outer_colour: [f32; 4],
         identification: [u16; 2],
         extra_data_1: [f32; 4],
-        extra_data_2: [f32; 4]) -> Self {
+        extra_data_2: [u16; 4]) -> Self {
         Self {
             position,
             dimensions,
@@ -55,7 +55,7 @@ impl RenderPrimitive {
             outer_colour,
             [0, 0],
             [stroke_width, 0.0, 0.0, 0.0],
-            [0.0, 0.0, 0.0, 0.0]
+            [0, 0, 0, 0]
         ) 
     }
 
@@ -65,7 +65,7 @@ impl RenderPrimitive {
         inner_colour: [f32; 4],
         outer_colour: [f32; 4],
         stroke_width: f32,
-        corner_radii: [f32; 4]) -> Self {
+        corner_radii: [u16; 4]) -> Self {
         RenderPrimitive::with_position_size_colours_identification_and_data(
             position,
             dimensions,
@@ -89,7 +89,7 @@ impl RenderPrimitive {
             colour,
             [2, glyph_index],
             [0.0, 0.0, 0.0, 0.0],
-            [0.0, 0.0, 0.0, 0.0]
+            [0, 0, 0, 0]
         ) 
     }
 }
